requirements.txt
Flask==2.3.3
Flask-RESTx==1.1.0
langchain==0.0.204
langchain-community==0.0.4
weaviate-client==3.27.1
transformers==4.31.0
PyMuPDF==1.23.1
torch==2.0.1
streamlit